moving detect multi channel in scpr_dataset
